+++ server/dashboard.js
@@ -208,11 +208,25 @@ async function downloadCsv(req, res) {
 
 async function renderDashboardPage(req, res) {
   try {
-    const displayedClients = (await db.getClients()).filter(client => client.isDisplayed)
+    const clients = (await db.getClients()).filter(client => client.isDisplayed)
 
-    const viewParams = {
-      clients: displayedClients,
-    }
+    const displayedClients = await Promise.all(
+      clients.map(async client => {
+        const clientExtension = await db.getClientExtensionWithClientId(client.id)
+        return {
+          ...client,
+          organization: clientExtension.organization || 'Unknown',
+        }
+      }),
+    )
+
+    displayedClients.sort((a, b) => {
+      const orgA = a.organization.toLowerCase()
+      const orgB = b.organization.toLowerCase()
+      return orgA.localeCompare(orgB)
+    })
+
+    const viewParams = { clients: displayedClients }
 
     res.send(Mustache.render(landingPageTemplate, viewParams, { nav: navPartial, css: landingCSSPartial }))
   } catch (err) {